% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ALL}
\alias{ALL}
\title{Ritz laboratory's ALL data}
\format{
An object of class \code{SummarizedExperiment} with 8585 rows and 125 columns.
}
\usage{
ALL
}
\description{
This is a clean and annotated version of the well-known \code{\link[ALL]{ALL}}
dataset from the Bioconductor package 'ALL' (which in its recent version 
doesn't contain feature annotations).
}
\references{
http://bioconductor.org/packages/release/data/experiment/html/ALL.html
}
\keyword{datasets}
