% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonomics.data-package.R
\docType{package}
\name{autonomics.data}
\alias{autonomics.data}
\alias{autonomics.data-package}
\title{Example omics datasets}
\description{
The package currently contains the following datasets:
\itemize{
  \item \code{\link{ALL}}:  autonomics-ready version of well-known ALL eset
  \item \code{billing2016}: stem cell differentiation data from Billing 2016 study
}
}
