% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-billing2016.R
\docType{data}
\name{billing2016}
\alias{billing2016}
\title{autonomics-ready eset with Billing 2016 data}
\format{An object of class \code{SummarizedExperiment} with 7877 rows and 9 columns.}
\usage{
billing2016
}
\description{
This is an autonomics-ready eset with the Billing 2016 data.
}
\references{
http://www.nature.com/articles/srep21507
}
\keyword{datasets}
