% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{glutaminase}
\alias{glutaminase}
\title{Halama et al. (2018) glutaminase inhibitor dataset}
\format{An object of class \code{SummarizedExperiment} with 391 rows and 96 columns.}
\usage{
glutaminase
}
\description{
Halama et al. (2018) glutaminase inhibitor dataset
}
\references{
https://doi.org/10.1016/j.canlet.2018.05.017
}
\keyword{datasets}
