% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stemcomp.proteinratios}
\alias{stemcomp.proteinratios}
\title{Billing et al. (2016) stem cell comparison: protein ratios (sample versus internal standard)}
\format{An object of class \code{SummarizedExperiment} with 7876 rows and 9 columns.}
\usage{
stemcomp.proteinratios
}
\description{
Billing et al. (2016) stem cell comparison: protein ratios (sample versus internal standard)
}
\references{
http://www.nature.com/articles/srep21507
}
\keyword{datasets}
