% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stemcomp.soma}
\alias{stemcomp.soma}
\title{Billing et al. (2016) stem cell comparison: somascan intensities}
\format{An object of class \code{SummarizedExperiment} with 1128 rows and 11 columns.}
\usage{
stemcomp.soma
}
\description{
Billing et al. (2016) stem cell comparison: somascan intensities
}
\references{
http://www.nature.com/articles/srep21507
}
\keyword{datasets}
